src/tag/requirements.txt
# NLP Core
transformers==4.45.0
tokenizers==0.20.0
docling-core==2.14.0      # ✅ Tanpa [chunking] untuk menghindari konflik

# LangChain
langchain==0.3.19
langchain-community==0.3.18
langchain-google-genai==2.1.0
langchain-openai==0.3.9
langchain-ollama==0.3.3
langchain-anthropic==0.3.3
langchain-huggingface==0.1.0
langchain-core== 0.3.64
langchain-chroma == 0.2.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Interface
pandas==2.2.3
selenium==4.28.1
dateparser==1.2.0

torch==2.2.2 --index-url https://download.pytorch.org/whl/cu124
torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cu124
torchaudio==2.2.2 --index-url https://download.pytorch.org/whl/cu124

ragas==0.2.14
jupyterlab==4.1.6
chainlit==2.5.5

huggingface-hub==0.25.2

matplotlib==3.10.3
rapidfuzz==3.13.0